backend/requirements.txt
fastapi[all]
pydantic
uvicorn
sqlalchemy
psycopg2-binary
requests